requirements.txt
django
djangorestframework 
drf-yasg
channels
channels_redis
daphne